synapseshield-ai/requirements.txt
fastapi
uvicorn[standard]
torch
scikit-learn
pandas
azure-identity
azure-digitaltwins-core
azure-eventhub
python-dotenv
